Add services, reorder functions to fit logic into controller. Add feed selector
