+++ index.js
@@ -1,7 +1,6 @@
 // require('./dist/main')
 
 const { app, BrowserWindow } = require('electron')
-const path = require('path')
 
 function createWindow () {
   const win = new BrowserWindow({
