+++ package.json
@@ -27,6 +27,7 @@
     "axios": "^0.27.2",
     "i18n": "^0.15.0",
     "m3u8-parser": "^6.0.0",
+    "m3u8stream": "^0.8.6",
     "prompts": "^2.4.2"
   }
 }
