+++ src/main/controllers/vod.controller.ts
@@ -1,0 +1,7 @@
+const getVod = async (url: string): Promise<void> => {
+  // Add Logic Here
+}
+
+export {
+  getVod
+}
