--- src/main/handlers/twitch.handler.ts
@@ -1,9 +1,0 @@
-// TODO: getAuth method from vod should be placed here
-
-function parseUrl (url: string): string {
-  const regex = /(?<=videos\/)(\d+)/g
-  const id = url.match(regex)
-  return id[0].toString()
-}
-
-exports.parseUrl = parseUrl
