+++ src/main/interfaces/Menu.ts
@@ -1,0 +1,3 @@
+export default interface Menu {
+  menu: string
+}
