+++ src/main/services/feed.service.ts
@@ -1,7 +1,7 @@
-import Playlist from '../../interfaces/Playlist'
-
+import Playlist from '../interfaces/Playlist'
 const m3u8Parser = require('m3u8-parser')
 
+// https://www.twitch.tv/videos/1629671834?filter=archives&sort=time
 // Passed a manifest string, returns an array of feeds
 const getFeedList = (manifest: string): object[] => {
   const parser = new m3u8Parser.Parser()
